LayoutTests/inspector/page/javascriptDialogEvents-expected.txt

CONFIRM NAVIGATION: beforeunload in javascriptDialogEvents
ALERT: alert
CONFIRM: confirm
PROMPT: prompt, default text: 
Opening with message: beforeunload in javascriptDialogEvents
Closed
Opening with message: alert
Closed
Opening with message: confirm
Closed
Opening with message: prompt
Closed
